v 0.3.0
major update. Fixed most of the potential bugs in previous update (v
0.3.0beta). Parallelization is used in multiple modules. Remaining
issues: 1. Floating point exception, which doesn’t appear in other
versions (v0.1.x and v 0.2.x). Issue appears in single- or multi-thread
modes. Cannot be reproduced even in the same thread. 2. The function to
compute AUC based on ooh samples is incorrect. In addition, the memory
consumption in multi-thread mode is needed to be further tested.