v0.5.0
major update. two options (--neighbor and --searchback) introduced in
v0.4.x are removed. they were supposed to increase power by exploring
the region of a marker. but it is suffered from the multiple-comparison
issue and has no benefit (as expected). after removing the options,
v0.5.0 is similar with v0.3.x but inherits minor updates and bug fixed
in v0.4.x