v 0.5.3
major update. In previous version, all covariates are selected randomly
as candidate splits. In general, covariates included in GWAS are
supposed to have significant effects, so always including them as the
candidates is more reason. This update adds --selcovar to randomly
select covariates as candidate splits, which is actually not
recommended in GWAS practice. By default, if this new option is off,
all covariates are treated as candidates in every split.